+++ wire-ios-request-strategy/Sources/Payloads/Payload+UpdateConversationStatus.swift
@@ -1,0 +1,59 @@
+//
+// Wire
+// Copyright (C) 2024 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import Foundation
+import WireDataModel
+
+extension Payload {
+    struct UpdateConversationStatus: Codable {
+        enum CodingKeys: String, CodingKey {
+            case mutedStatus = "otr_muted_status"
+            case mutedReference = "otr_muted_ref"
+            case archived = "otr_archived"
+            case archivedReference = "otr_archived_ref"
+            case hidden = "otr_hidden"
+            case hiddenReference = "otr_hidden_ref"
+        }
+
+        var mutedStatus: Int?
+        var mutedReference: Date?
+        var archived: Bool?
+        var archivedReference: Date?
+        var hidden: Bool?
+        var hiddenReference: String?
+
+        init(_ conversation: ZMConversation) {
+
+            if conversation.hasLocalModifications(forKey: ZMConversationSilencedChangedTimeStampKey) {
+                let reference = conversation.silencedChangedTimestamp ?? Date()
+                conversation.silencedChangedTimestamp = reference
+
+                mutedStatus = Int(conversation.mutedMessageTypes.rawValue)
+                mutedReference = reference
+            }
+
+            if conversation.hasLocalModifications(forKey: ZMConversationArchivedChangedTimeStampKey) {
+                let reference = conversation.archivedChangedTimestamp ?? Date()
+                conversation.archivedChangedTimestamp = reference
+
+                archived = conversation.isArchived
+                archivedReference = reference
+            }
+        }
+    }
+}
